fix (budget sheet): Update table logic to display proper data
Each heading displays different data based on whether the user looks in the past, present, or future.
In past and present months the table shows the budget and spendings of that month, the user will need to cover spendings
with a budget or available cash will turn red. In a surplus budget case for past months the money will roll into present and future months.
The future does not show any spendings as it did not happen yet. In future months the budget  allows the user to see a categorie's budget
from the previous months, but does not include the categorie's budget from the future months.
The only heading that doesn't change is the Pre-Set heading as it is intended to be a hardcoded baseline for the user to set.
Available Cash does not change based on the view date and is the amount of money in the account that is not allocated to a specfic category.